Extracted classes from Main.java to separated files.
